requirements.txt
aiohttp>=3.9
aiosqlite>=0.19
fastapi>=0.109
uvicorn[standard]>=0.27
websockets>=12.0
mss>=9.0
Pillow>=10.2
pyyaml>=6.0
pydantic>=2.5
httpx>=0.26
apscheduler>=3.10
rich>=13.7
